+++ esp-hal/src/clock/clocks_ll/esp32c6.rs
@@ -2,7 +2,6 @@ use crate::{
     clock::{ApbClock, Clock, CpuClock, PllClock, XtalClock},
     peripherals::{MODEM_LPCON, MODEM_SYSCON, PMU},
     rtc_cntl::rtc::CpuClockSource,
-    system::{RadioClockController, RadioPeripherals},
 };
 
 const I2C_BBPLL: u8 = 0x66;
@@ -450,43 +449,7 @@ pub(crate) fn esp32c6_bbpll_get_freq_mhz() -> u32 {
     CLK_LL_PLL_480M_FREQ_MHZ
 }
 
-impl RadioClockController for crate::peripherals::RADIO_CLK {
-    fn enable(&mut self, peripheral: RadioPeripherals) {
-        match peripheral {
-            RadioPeripherals::Phy => enable_phy(true),
-            RadioPeripherals::Wifi => wifi_clock_enable(true),
-            RadioPeripherals::Bt => ble_clock_enable(true),
-            RadioPeripherals::Ieee802154 => ieee802154_clock_enable(true),
-        }
-    }
-
-    fn disable(&mut self, peripheral: RadioPeripherals) {
-        match peripheral {
-            RadioPeripherals::Phy => enable_phy(false),
-            RadioPeripherals::Wifi => wifi_clock_enable(false),
-            RadioPeripherals::Bt => ble_clock_enable(false),
-            RadioPeripherals::Ieee802154 => ieee802154_clock_enable(false),
-        }
-    }
-
-    fn reset_mac(&mut self) {
-        reset_mac();
-    }
-
-    fn init_clocks(&mut self) {
-        init_clocks();
-    }
-
-    fn ble_rtc_clk_init(&mut self) {
-        // nothing for this target (yet)
-    }
-
-    fn reset_rpa(&mut self) {
-        // nothing for this target (yet)
-    }
-}
-
-fn enable_phy(en: bool) {
+pub(super) fn enable_phy(en: bool) {
     MODEM_LPCON::regs()
         .clk_conf()
         .modify(|_, w| w.clk_i2c_mst_en().bit(en));
@@ -495,7 +458,7 @@ fn enable_phy(en: bool) {
         .modify(|_, w| w.clk_i2c_mst_sel_160m().bit(en));
 }
 
-fn wifi_clock_enable(en: bool) {
+pub(super) fn enable_wifi(en: bool) {
     MODEM_SYSCON::regs().clk_conf1().modify(|_, w| {
         w.clk_wifi_apb_en().bit(en);
         w.clk_wifimac_en().bit(en);
@@ -520,7 +483,7 @@ fn wifi_clock_enable(en: bool) {
     });
 }
 
-fn ieee802154_clock_enable(en: bool) {
+pub(super) fn enable_ieee802154(en: bool) {
     MODEM_SYSCON::regs().clk_conf().modify(|_, w| {
         w.clk_zb_apb_en().bit(en);
         w.clk_zb_mac_en().bit(en)
@@ -549,7 +512,7 @@ fn ieee802154_clock_enable(en: bool) {
         .modify(|_, w| w.clk_coex_en().set_bit());
 }
 
-fn ble_clock_enable(en: bool) {
+pub(super) fn enable_bt(en: bool) {
     MODEM_SYSCON::regs().clk_conf().modify(|_, w| {
         w.clk_etm_en().bit(en);
         w.clk_modem_sec_en().bit(en);
@@ -574,11 +537,11 @@ fn ble_clock_enable(en: bool) {
         .modify(|_, w| w.clk_coex_en().bit(en));
 }
 
-fn reset_mac() {
+pub(super) fn reset_mac() {
     // empty
 }
 
-fn init_clocks() {
+pub(super) fn init_clocks() {
     unsafe {
         let pmu = PMU::regs();
 
@@ -638,4 +601,12 @@ fn init_clocks() {
             .clk_conf()
             .modify(|_, w| w.clk_wifipwr_en().set_bit());
     }
+}
+
+pub(super) fn ble_rtc_clk_init() {
+    // nothing for this target (yet)
+}
+
+pub(super) fn reset_rpa() {
+    // nothing for this target (yet)
 }