+++ esp-hal/src/clock/clocks_ll/esp32s2.rs
@@ -1,8 +1,4 @@
-use crate::{
-    clock::CpuClock,
-    peripherals::SYSCON,
-    system::{RadioClockController, RadioPeripherals},
-};
+use crate::{clock::CpuClock, peripherals::SYSCON};
 
 const MHZ: u32 = 1000000;
 const UINT16_MAX: u32 = 0xffff;
@@ -51,67 +47,31 @@ pub(crate) fn set_cpu_clock(cpu_clock_speed: CpuClock) {
 const DPORT_WIFI_CLK_WIFI_BT_COMMON_M: u32 = 0x000003c9;
 const DPORT_WIFI_CLK_WIFI_EN_M: u32 = 0x000007cf;
 
-impl RadioClockController for crate::peripherals::RADIO_CLK {
-    fn enable(&mut self, peripheral: RadioPeripherals) {
-        match peripheral {
-            RadioPeripherals::Phy => enable_phy(),
-            RadioPeripherals::Wifi => wifi_clock_enable(),
-        }
-    }
-
-    fn disable(&mut self, peripheral: RadioPeripherals) {
-        match peripheral {
-            RadioPeripherals::Phy => disable_phy(),
-            RadioPeripherals::Wifi => wifi_clock_disable(),
-        }
-    }
-
-    fn reset_mac(&mut self) {
-        reset_mac();
-    }
-
-    fn init_clocks(&mut self) {
-        init_clocks();
-    }
-
-    fn ble_rtc_clk_init(&mut self) {
-        // nothing for this target
-    }
-
-    fn reset_rpa(&mut self) {
-        // nothing for this target
-    }
-}
-
-fn enable_phy() {
+pub(super) fn enable_phy(enable: bool) {
     // `periph_ll_wifi_bt_module_enable_clk_clear_rst`
-    SYSCON::regs()
-        .wifi_clk_en()
-        .modify(|r, w| unsafe { w.bits(r.bits() | DPORT_WIFI_CLK_WIFI_BT_COMMON_M) });
-}
-
-fn disable_phy() {
     // `periph_ll_wifi_bt_module_disable_clk_set_rst`
-    SYSCON::regs()
-        .wifi_clk_en()
-        .modify(|r, w| unsafe { w.bits(r.bits() & !DPORT_WIFI_CLK_WIFI_BT_COMMON_M) });
+    SYSCON::regs().wifi_clk_en().modify(|r, w| unsafe {
+        if enable {
+            w.bits(r.bits() | DPORT_WIFI_CLK_WIFI_BT_COMMON_M)
+        } else {
+            w.bits(r.bits() & !DPORT_WIFI_CLK_WIFI_BT_COMMON_M)
+        }
+    });
 }
 
-fn wifi_clock_enable() {
+pub(super) fn enable_wifi(enable: bool) {
     // `periph_ll_wifi_module_enable_clk_clear_rst`
-    SYSCON::regs()
-        .wifi_clk_en()
-        .modify(|r, w| unsafe { w.bits(r.bits() | DPORT_WIFI_CLK_WIFI_EN_M) });
-}
-
-fn wifi_clock_disable() {
     // `periph_ll_wifi_module_disable_clk_set_rst`
-    SYSCON::regs()
-        .wifi_clk_en()
-        .modify(|r, w| unsafe { w.bits(r.bits() & !DPORT_WIFI_CLK_WIFI_EN_M) });
+    SYSCON::regs().wifi_clk_en().modify(|r, w| unsafe {
+        if enable {
+            w.bits(r.bits() | DPORT_WIFI_CLK_WIFI_EN_M)
+        } else {
+            w.bits(r.bits() & !DPORT_WIFI_CLK_WIFI_EN_M)
+        }
+    });
 }
 
-fn reset_mac() {
+pub(super) fn reset_mac() {
     const SYSTEM_MAC_RST: u32 = 1 << 2;
     SYSCON::regs()
         .wifi_rst_en()
@@ -121,7 +81,7 @@ fn reset_mac() {
         .modify(|r, w| unsafe { w.wifi_rst().bits(r.wifi_rst().bits() & !SYSTEM_MAC_RST) });
 }
 
-fn init_clocks() {
+pub(super) fn init_clocks() {
     const DPORT_WIFI_CLK_WIFI_EN: u32 = 0x003807cf;
     const DPORT_WIFI_CLK_BT_EN_M: u32 = 0x61 << 11;
     const DPORT_WIFI_CLK_SDIOSLAVE_EN: u32 = 1 << 4;
@@ -141,4 +101,12 @@ fn init_clocks() {
     SYSCON::regs()
         .wifi_clk_en()
         .modify(|r, w| unsafe { w.bits(r.bits() & !WIFI_BT_SDIO_CLK | DPORT_WIFI_CLK_WIFI_EN) });
+}
+
+pub(super) fn ble_rtc_clk_init() {
+    // nothing for this target
+}
+
+pub(super) fn reset_rpa() {
+    // nothing for this target
 }