+++ esp-hal/src/clock/clocks_ll/esp32s3.rs
@@ -2,7 +2,6 @@ use crate::{
     clock::{Clock, CpuClock},
     peripherals::APB_CTRL,
     rom,
-    system::{RadioClockController, RadioPeripherals},
 };
 
 pub(crate) fn set_cpu_clock(cpu_clock_speed: CpuClock) {
@@ -35,63 +34,29 @@ const SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M: u32 = 0x78078F;
 // SYSTEM_WIFI_CLK_EN : R/W ;bitpos:[31:0] ;default: 32'hfffce030
 const SYSTEM_WIFI_CLK_EN: u32 = 0x00FB9FCF;
 
-impl RadioClockController for crate::peripherals::RADIO_CLK {
-    fn enable(&mut self, peripheral: RadioPeripherals) {
-        match peripheral {
-            RadioPeripherals::Phy => enable_phy(),
-            RadioPeripherals::Bt => common_wifi_bt_clock_enable(),
-            RadioPeripherals::Wifi => common_wifi_bt_clock_enable(),
-        }
-    }
-
-    fn disable(&mut self, peripheral: RadioPeripherals) {
-        match peripheral {
-            RadioPeripherals::Phy => disable_phy(),
-            RadioPeripherals::Bt => common_wifi_bt_clock_disable(),
-            RadioPeripherals::Wifi => common_wifi_bt_clock_disable(),
-        }
-    }
-
-    fn reset_mac(&mut self) {
-        reset_mac();
-    }
-
-    fn init_clocks(&mut self) {
-        init_clocks();
-    }
-
-    fn ble_rtc_clk_init(&mut self) {
-        // nothing for this target
-    }
-
-    fn reset_rpa(&mut self) {
-        // nothing for this target
-    }
-}
-
-fn enable_phy() {
+pub(super) fn enable_phy(enable: bool) {
     // `periph_ll_wifi_bt_module_enable_clk_clear_rst`
-    APB_CTRL::regs()
-        .wifi_clk_en()
-        .modify(|r, w| unsafe { w.bits(r.bits() | SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M) });
-}
-
-fn disable_phy() {
     // `periph_ll_wifi_bt_module_disable_clk_set_rst`
-    APB_CTRL::regs()
-        .wifi_clk_en()
-        .modify(|r, w| unsafe { w.bits(r.bits() & !SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M) });
+    APB_CTRL::regs().wifi_clk_en().modify(|r, w| unsafe {
+        if enable {
+            w.bits(r.bits() | SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M)
+        } else {
+            w.bits(r.bits() & !SYSTEM_WIFI_CLK_WIFI_BT_COMMON_M)
+        }
+    });
 }
 
-fn common_wifi_bt_clock_enable() {
+pub(super) fn enable_bt(_: bool) {
     // `periph_ll_wifi_module_enable_clk_clear_rst`. does nothing
-}
-
-fn common_wifi_bt_clock_disable() {
     // `periph_ll_wifi_module_disable_clk_set_rst`. does nothing
 }
 
-fn reset_mac() {
+pub(super) fn enable_wifi(_: bool) {
+    // `periph_ll_wifi_module_enable_clk_clear_rst`. does nothing
+    // `periph_ll_wifi_module_disable_clk_set_rst`. does nothing
+}
+
+pub(super) fn reset_mac() {
     const SYSTEM_MAC_RST: u32 = 1 << 2;
     APB_CTRL::regs()
         .wifi_rst_en()
@@ -101,7 +66,7 @@ fn reset_mac() {
         .modify(|r, w| unsafe { w.bits(r.bits() & !SYSTEM_MAC_RST) });
 }
 
-fn init_clocks() {
+pub(super) fn init_clocks() {
     const SYSTEM_WIFI_CLK_I2C_CLK_EN: u32 = 1 << 5;
     const SYSTEM_WIFI_CLK_UNUSED_BIT12: u32 = 1 << 12;
     const SYSTEM_WIFI_CLK_SDIO_HOST_EN: u32 = 1 << 13;
@@ -112,4 +77,12 @@ fn init_clocks() {
     APB_CTRL::regs()
         .wifi_clk_en()
         .modify(|r, w| unsafe { w.bits(r.bits() & !WIFI_BT_SDIO_CLK | SYSTEM_WIFI_CLK_EN) });
+}
+
+pub(super) fn ble_rtc_clk_init() {
+    // nothing for this target
+}
+
+pub(super) fn reset_rpa() {
+    // nothing for this target
 }