+++ esp-wifi/src/common_adapter/mod.rs
@@ -1,10 +1,9 @@
 use esp_wifi_sys::include::timeval;
-use hal::ram;
 
 use crate::{
     binary::include::{esp_event_base_t, esp_timer_get_time},
     compat::{common::*, timer_compat::*},
-    hal,
+    hal::{self, clock::RadioClockController, peripherals::RADIO_CLK, ram},
 };
 
 #[cfg_attr(esp32c3, path = "common_adapter_esp32c3.rs")]
@@ -114,7 +113,7 @@ pub unsafe extern "C" fn random() -> crate::binary::c_types::c_ulong {
     trace!("random");
 
     // stealing RNG is safe since we own it (passed into `init`)
-    let mut rng = esp_hal::rng::Rng::new(unsafe { esp_hal::peripherals::RNG::steal() });
+    let mut rng = hal::rng::Rng::new(unsafe { hal::peripherals::RNG::steal() });
     rng.random()
 }
 
@@ -135,7 +134,7 @@ pub unsafe extern "C" fn random() -> crate::binary::c_types::c_ulong {
 pub unsafe extern "C" fn read_mac(mac: *mut u8, type_: u32) -> crate::binary::c_types::c_int {
     trace!("read_mac {:?} {}", mac, type_);
 
-    let base_mac = crate::hal::efuse::Efuse::mac_address();
+    let base_mac = hal::efuse::Efuse::mac_address();
 
     for (i, &byte) in base_mac.iter().enumerate() {
         mac.add(i).write_volatile(byte);
@@ -306,4 +305,23 @@ pub unsafe extern "C" fn strrchr(_s: *const (), _c: u32) -> *const u8 {
 #[no_mangle]
 pub unsafe extern "C" fn floor(v: f64) -> f64 {
     libm::floor(v)
+}
+
+// TODO: previously ESP32 and S2 refcounted the PHY clock. Should we?
+
+pub(crate) unsafe fn phy_enable_clock() {
+    // stealing RADIO_CLK is safe since it is passed (as mutable reference or by
+    // value) into `init`
+    let radio_clocks = unsafe { RADIO_CLK::steal() };
+    RadioClockController::new(radio_clocks).enable_phy(true);
+    trace!("phy_enable_clock done!");
+}
+
+#[allow(unused)]
+pub(crate) unsafe fn phy_disable_clock() {
+    // stealing RADIO_CLK is safe since it is passed (as mutable reference or by
+    // value) into `init`
+    let radio_clocks = unsafe { RADIO_CLK::steal() };
+    RadioClockController::new(radio_clocks).enable_phy(false);
+    trace!("phy_disable_clock done!");
 }